--- docs/index.html
@@ -1,113 +1,0 @@
-<!doctype html>
-    <html>
-    <head>
-        <title>Qdox</title>
-        <meta name="expires" content="86400" />
-        <link rel="stylesheet" href="qdox.css">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-        <link rel="stylesheet" href="https://www.w3schools.com/w3css/4/w3.css">
-    </head>
-    <body>
-
-    <!-- Sidebar -->
-    <div class="w3-sidebar w3-light-grey w3-bar-block" style="width:180px">
-      <center>
-        <img src="https://avatars.githubusercontent.com/u/21315775?v=4" height="128px" width="128px"/>
-        <br/><a href="https://github.com/eudoxys">Eudoxys Sciences LLC</a>
-        <br/><a href="https://github.com/eudoxys"></a>
-      </center>
-      <title class="w3-bar-item">qdox</title>
-      <a href="#main" class="w3-bar-item w3-button">Command Line</a>
-      <a href="#python" class="w3-bar-item w3-button">Python Library</a>
-      <a href="#package" class="w3-bar-item w3-button">Package Metadata</a>
-    </div>
-
-    <!-- Page Content -->
-    <div style="margin-left:220px">
-
-    <h1 id="main" class="w3-container">Command Line</h1>
-
-    <p/>
-Generate docs/index.html from module contents and metadata
-<p/>Syntax: <code>qdox [OPTION ...]</code><p/>
-<h2 class="w3-container">Options</h2>
-<p/><ul><li>
-<code>-</code>: run with no options</li><li>
-<code>--debug</code>: Enable debugging traceback on exceptions</li><li>
-<code>-h|--help|help</code>: Display this help information</li><li>
-<code>--tomlfile=TOMFILE</code>: Use TOMFILE instead of "pyproject.toml"</li><li>
-<code>--withcss[=CSSFILE]</code>: Copy the CSS file to <code>docs/</code> (default is <code>qdox.css</code>)</li></li>
-</ul>
-
-<h2 class="w3-container">Description</h2>
-<p/>  The <code>qdox</code> command generates the documentation for a simple Python
-  project. The formatting is designed to use simple text layout as the input 
-  so that the same documentation source can be used for both Python <code>help()</code> 
-  output and the documentation pages.
-<p/>  The command loads the module specified in the <code>pyproject.toml</code> file and
-  outputs the <code>__doc__</code> property of the module as the command line
-  documentation. It then outputs the python functions and constants,
-  followed by the project metadata.
-<p/>  You can use your project's <code>pages-build-deployment</code> workflow to deploy the
-  documentation to <code>github.io</code> or to your own custom site by setting the
-  following options in your project <code>Settings-->Pages</code>:
-<p/><ul><li>
-<code><code>Source</code></code>: Choose <b>Deploy from a branch</b></li><li>
-<code><code>Branch</code></code>: Choose <b>main</b> and <b>/docs</b>.</li><li>
-<code><code>Custom domain</code></code>: Enter one if you have one.</li></li>
-</ul>
-
-<h2 class="w3-container">Usage</h2>
-<p/>  To use <code>qdox</code> in a project you must do the following.
-<p/><ol><li> Install <code>qdox</code> from the repo into the current <code>venv</code> by running the</li>  command   <pre>python3 -m pip install git+<a href="https://github.com/dchassin/qdox" target="_tab">https://github.com/dchassin/qdox</a></pre>   <li> Run <code>qdox</code> in the project root folder by running the command</li>  <pre>qdox -</pre>    If you want to include a CSS file, use the option <code>--withcss=CSSFILE</code>, e.g.,  <pre>qdox --withcss=mystyles.css</pre>  The <code>CSSFILE</code> defaults to <code>qdox.css</code> when omitted.  <li> Add/push the new files to GitHub by running the commands </li>  <pre>git commit -a -m "Add docs"git push</pre></li>
-</ol>
-
-<h2 class="w3-container">Text Formatting</h2>
-<p/>  Text encoding includes the following
-<p/><ul><li>
-<code><code>Syntax</code>: text</code>: Syntax text</li><li>
-<code><code>text:</code></code>: Level 2 heading text in module docs or level 3 in function docs</li><li>
-<code><code>(blank line)</code></code>: New paragraph</li><li>
-<code><code>(four spaces)text</code></code>: Preformatted text</li><li>
-<code><code>(four spaces)*(space)text</code></code>: Bulleted list text</li><li>
-<code><code>(four spaces)(digit)(space)text</code></code>: Numbered list item</li><li>
-<code><code>(seven spaces)text</code></code>: List item text continuation</li><li>
-<code><code>(nine spaces)text</code></code>: Preformatted text in a list item</li><li>
-<code><code>**text**</code></code>: <b>Bold</b> text</li><li>
-<code><code>*text*</code></code>: <i>Italic</i> text<pre>
-</pre></li><li>
-<code><code>!!text!!</code></code>: <font class="highlight">Highlight</font> text</li><li>
-<code><code>~~text~~</code></code>: <strike>strikeout</strike> text</li><li>
-<code><code>__text__</code></code>: <u>underline</u> text</li><li>
-<code><code>^word</code></code>: <sup>superscript</sup> word (ends at space)<pre>
-</pre></li><li>
-<code><code>_word</code></code>: <sub>subscript</sub> word (ends at space)</li><li>
-<code>`text`</code>: <code>code</code> text</li><li>
-<code><code>protocol://url/</code></code>: <a href="protocol://url/" target="_tab">protocol://url/</a> formatting with active link</li></li>
-</ul>
-
-<h2 class="w3-container">Pro-tips</h2>
-<p/><pre>The module you are documenting must be installed in the active Python
-environment for <code>qdox</code> to read information about the module. 
-<p/>If you want to output a colon at the end of a paragraph rather than a
-heading, place a space after the colon and before the end-of-line.
-<p/></pre>
-<h1 id="python" class="w3-container">Python Library</h1>
-
-<h1 id="package" class="w3-container">Package Metadata</h1>
-<p/><table class="w3-container">
-<tr><th><nobr>Name</nobr></th><td>:</td><td>qdox</td></tr>
-<tr><th><nobr>Version</nobr></th><td>:</td><td>0.0.0a4</td></tr>
-<tr><th><nobr>Description</nobr></th><td>:</td><td>Generate quick documentation for a Python project on GitHub</td></tr>
-<tr><th><nobr>Authors</nobr></th><td>:</td><td>David P. Chassin <dpchassin@gmail.com></td></tr>
-<tr><th><nobr>Maintainers</nobr></th><td>:</td><td>David P. Chassin <dpchassin@gmail.com></td></tr>
-<tr><th><nobr>Requires-Python</nobr></th><td>:</td><td>>= 3.11</td></tr>
-<tr><th><nobr>Dependencies</nobr></th><td>:</td><td>requests<br/>tomli</td></tr>
-<tr><th><nobr>Keywords</nobr></th><td>:</td><td>github<br/>python<br/>documentation</td></tr>
-<tr><th><nobr>License</nobr></th><td>:</td><td>MIT License</td></tr>
-<tr><th><nobr>Classifiers</nobr></th><td>:</td><td>Development Status :: 3 - Alpha<br/>Intended Audience :: Developers<br/>License :: OSI Approved :: MIT License<br/>Programming Language :: Python :: 3 :: Only<br/>Topic :: Software Development :: Libraries</td></tr>
-<tr><th><nobr>Urls</nobr></th><td>:</td><td>Homepage = <a href="https://github.com/eudoxys/qdox" target="_tab">https://github.com/eudoxys/qdox</a><br/>Documentation = <a href="https://www.eudoxys.com/qdox" target="_tab">https://www.eudoxys.com/qdox</a><br/>Repository = <a href="https://github.com/eudoxys/qdox.git" target="_tab">https://github.com/eudoxys/qdox.git</a><br/>Issues = <a href="https://github.com/eudoxys/qdox/issues" target="_tab">https://github.com/eudoxys/qdox/issues</a></td></tr>
-<tr><th><nobr>Scripts</nobr></th><td>:</td><td><code>qdox</code> &rightarrow; <code>main()</code></td></tr>
-</table>
-<hr/><p/><cite>Copyright &copy; 2024 David P. Chassin</body>
-            </html>
