--- docs/qdox.css
@@ -1,66 +1,0 @@
-title {
-    color:#fff!important;
-    background-color:#009688!important
-}
-
-h1 {
-    color:#fff!important;
-    background-color:#009688!important
-}
-
-h2 {
-    background-color: lightgrey;
-    color: black;
-}
-
-h3 {
-    text-decoration: underline;
-}
-
-li {
-    margin-top: 5px;
-}
-
-table {
-    border-spacing: 10px;
-}
-
-th {
-    text-align: left;
-}
-
-tr {
-    vertical-align: top;
-}
-
-pre {
-    font-family: courier;
-    background-color: #eeeeee;
-    margin-left: 20px;
-}
-
-p {
-    margin-left: 20px;
-    margin-right: 40px;
-}
-
-font.highlight
-{
-    background-color: lightyellow;
-}
-
-a:hover {
-    background-color: lightyellow;
-}
-
-code {
-    font-family: courier;
-}
-
-code:hover {
-    background-color: lightblue;
-}
-
-pre:hover {
-    background-color: lightblue;
-}
